+++ changelog.md
@@ -4,6 +4,8 @@ Saves from 8.x are not compatible with 9.0.0.
 
 ## Features/Improvements
 
+* **[Modding]** Factions can now specify the ship type to be used for cargo shipping. The Handy Wind will be used by default, but WW2 factions can pick something more appropriate.
+
 ## Fixes
 
 # 8.1.0